adjust types + reformat sql migraitons

+++ internal/database/migrations/20231204162013_create_books.up.sql
@@ -3,22 +3,17 @@ CREATE TABLE
     id serial PRIMARY KEY,
     title text NOT NULL DEFAULT '',
     ISBN text NOT NULL,
-    price integer NOT NULL,
+    price numeric(14,2) NOT NULL,
     publisher text NOT NULL,
     pages integer NOT NULL,
     publish_date date NOT NULL,
-    created_at timestamp(0)
-    with
-      time zone NOT NULL DEFAULT now (),
-      updated_at timestamp(0)
-    with
-      time zone NOT NULL DEFAULT now (),
-      author_id integer NOT NULL,
-      CONSTRAINT books_author_id_fk FOREIGN KEY (author_id) REFERENCES authors (id)
+    created_at timestamp(0) with time zone NOT NULL DEFAULT now (),
+    updated_at timestamp(0) with time zone NOT NULL DEFAULT now (),
+    author_id integer NOT NULL,
+    CONSTRAINT books_author_id_fk FOREIGN KEY (author_id) REFERENCES authors (id)
   );
 
 CREATE INDEX IF NOT EXISTS idx_books_title ON books (title);
 
 CREATE UNIQUE INDEX IF NOT EXISTS idx_books_isbn ON books (ISBN)
-WHERE
-  ISBN <> '';
+WHERE ISBN <> '';
